fix: refactor and colorutility implementation
