+++ packages/react/src/themes/default-theme/index.ts
@@ -1,2 +1,4 @@
 export * from './default-theme';
+export * from './default-ref-tokens';
 export * from './default-component-tokens';
+export * from './default-legacy-tokens';
