+++ packages/react/src/themes/tokens/index.ts
@@ -1,5 +1,5 @@
-export * from './palette';
-export * from './text-attributes';
+export * from './ref-tokens/palette';
+export * from './ref-tokens/text-attributes';
 export * from './ref-tokens';
 export * from './alias-tokens';
 export * from './component-tokens';
