+++ packages/react/src/themes/tokens/ref-tokens.ts
@@ -1,9 +1,16 @@
-import { PaletteTokens, PaletteValue } from './palette';
-import { TextAttributeTokens, TextAttributeValue } from './text-attributes';
+import { ColorUtilityTokens, ColorUtilityValue } from './ref-tokens/color-utility-tokens';
+import { PaletteTokens, PaletteValue } from './ref-tokens/palette-tokens';
+import { TextAttributeTokens, TextAttributeValue } from './ref-tokens/text-attributes-tokens';
 
-export type RefTokenValue = PaletteValue | TextAttributeValue;
+export type RefTokenValue =
+    | PaletteValue
+    | TextAttributeValue
+    | ColorUtilityValue;
 
-export type RefTokens = PaletteTokens | TextAttributeTokens;
+export type RefTokens =
+    | PaletteTokens
+    | TextAttributeTokens
+    | ColorUtilityTokens;
 
 export type RefTokenMap = {
     [Token in RefTokens]: RefTokenValue;
