+++ packages/react/src/themes/tokens/ref-tokens/text-attributes-tokens.ts
@@ -1,8 +1,7 @@
 export type TextAttributeValue = string;
 
 export type TextAttributeTokens =
-    | 'font-weight-bold'
-    | 'transparent';
+    | 'font-weight-bold';
 
 export type TextAttributeTokenMap = {
     [Token in TextAttributeTokens]: TextAttributeValue;
